+++ app/src/main/java/project/paveltoy/movietap/ui/favorites/FavoriteMovieFragment.kt
@@ -12,6 +12,10 @@ import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.snackbar.BaseTransientBottomBar
 import com.google.android.material.snackbar.Snackbar
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.launch
 import project.paveltoy.movietap.R
 import project.paveltoy.movietap.data.entity.MovieEntity
 import project.paveltoy.movietap.data.getTextForIsFavoriteSnackbar
@@ -54,7 +58,9 @@ class FavoriteMovieFragment : Fragment() {
     }
 
     private fun setAdapter() {
-        viewModel.getFavoriteMovies()
+        CoroutineScope(Dispatchers.Main).launch {
+            viewModel.getFavoriteMovies()
+        }
         viewModel.favoriteMovies.observe(viewLifecycleOwner) {
             adapter.data = it
             adapter.notifyDataSetChanged()
@@ -73,7 +79,9 @@ class FavoriteMovieFragment : Fragment() {
     private fun setOnFavoriteChanged() {
         adapter.onFavoriteChanged = { movie: MovieEntity, index: Int ->
             movie.isFavorite = !movie.isFavorite
-            viewModel.removeFromFavorite(movie)
+            CoroutineScope(Dispatchers.Main).launch {
+                viewModel.removeFromFavorite(movie)
+            }
             val text = getTextForIsFavoriteSnackbar(resources, movie.title, movie.isFavorite)
             Snackbar.make(requireView(), text, BaseTransientBottomBar.LENGTH_LONG)
                 .setAnchorView(R.id.bottom_navigation)
